allow particles of child particle systems to be launched with (dynamic) normals
new particle script constructs (in particle blocks):
  {velocityType | accelerationType} {last_normal | opportunistic_normal}

the base vector in case of opportunistic_normal is (0,0,0) if the particle is flying, and is a surface normal vector if the particle is in/on a map surface. the base vector in case of last_normal is always a valid direction, though sometimes an arbitrary one (the up direction). in both cases, the initial value of the last/opportunistic normal vector is propagated from the parent particle.

the last/opportunistic vector is a new particle system variable, and the original normal vector is not touched, so this change retains backward-compatibility.