Create home path if it doesn't exist.
This commit touches the download file to make sure it can be written to
prior to retrieving the remote files.

Home path won't exist on a fresh installation, and the restclient code
does not uses the built FS_ utilities to write open/write the new files
making this a required step.