Check for truncated paths in Sys_LoadDll
Check for truncated paths which could allow loading a library with a
non-standard extension. Also provides a better message for why a valid
library with a long path would fail to load.

Don't load libraries with non-standard file extensions

Also don't allow writting files ending in a library extension such as
".so.0" or ".dylib.0".