Client: Avoid unwind crashes when long jumping out of the VM error handler.
On Windows 64 bit, the VMI_COMPILED (vm_x86.cpp) code generates
code which is not compatible with the unwind handler. In 32bit this
isn't an issue (perhaps call convention).

This is a large outstanding issue after the C++ conversion for Windows.
In the Microsoft world, longjmp() will unwind and call deconstructors on
it's way back to setjmp(). It's not obvious if this happens on Linux and
OSX as well due to vague documentation, but I expect it has more to do
with the code generation in vm_x86.cpp not matching the 64bit FASTCALL
ABI.

GH-44